backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.6.1
motor==3.6.0
python-dotenv==1.0.1

# Optional heavy dependencies (enable when USE_STUB=0)
# torch==2.3.1
# openai-whisper==20231117
# transformers==4.44.2
# sentencepiece==0.2.0
# pyannote.audio==3.3.1
